generate_latency_report.py: add iteration count
An iteration counter is now tracked by sv_timestamp_logger.
This counter start at zero and increase every time the SV counter go
back to 0.
It allows tracking every SV with a unique identifier.

This commit modifies the latency report script to handle this modification
and to display the iteration counter.

Signed-off-by: Paul Le Guen de Kerneizon <paul.leguendekerneizon@savoirfairelinux.com>