scripts: correct time global variable
The current implementation of the script uses a global variable to log
the time of entry of the SV in the machine. It then compares it to the
time when the SV is available in userspace.
A problem occurs when multiple SV arrives at the same time.

This new version uses the map function of bpftrace which logs one
different entry time for each SV.

Signed-off-by: Paul Le Guen de Kerneizon <paul.leguendekerneizon@savoirfairelinux.com>