scripts: duplicate sv_iteration variable
The current version of the bpftrace scripts uses the sv_counter and the
sv_iteration as indexes to register the time of a SV at the entry of the
script.

However, we get the time of an SV at two different functions (entry and
exit) and sv_iteration is a global variable in the script.
Thus, the sv_iteration variable must be declared and handled seperatly
in the two functions to correctly target each SV.

Signed-off-by: Paul Le Guen de Kerneizon <paul.leguendekerneizon@savoirfairelinux.com>