generate_latency_report.py: plot seapath time
The goal of this tool is
- to measure the latency inside the hypervisor's kernel
- to measure the latency inside the VM's kernel
- to measure the latency to pass a packet from the hypervisor to the VM.

It then seems not relevant to print the latency of the network.

This commit uses the hardware timestamp taken on the NIC of the
hypervisor to calculate the "SEAPATH latency" : from the entry of the
NIC of the hypervisor to the VM in userspace.
It also removes all mentions of the Tres/Tnet that was used to calculate
the network time.

Signed-off-by: Paul Le Guen de Kerneizon <paul.leguendekerneizon@savoirfairelinux.com>