launch_bpftrace: set BPFTRACE_MAX_MAP_KEYS to 16384
By default, BPFtrace map can contains a maximum of 4096 keys. When
running tests with many streams, this maximum can be easily achieved,
when a lot of SV are incoming of the SV interface and not proceed yet
(and flushed from the map).

To prevent this, set this maximum to 16384 keys.

Signed-off-by: Paul Le Guen de Kerneizon <paul.leguendekerneizon@savoirfairelinux.com>